+++ src/memory.h
@@ -16,8 +16,8 @@
 //Fetches an instruction, decodes and updates current_inst
 void Fetch_Decode(instruction & current_inst, bool I_or_D);
 
-//Perform a data fetch  
-int access_mem(int index, int flag, int out=0); 
+//Perform a data fetch
+int access_mem(uint16_t index, int flag, int out=0);
 
 
 #endif//MEMORY_H_
